DATA-1932 Revert ItemStatus name change
