+++ src/main/java/com/plaid/client/response/IdentityGetResponse.java
@@ -3,10 +3,10 @@ package com.plaid.client.response;
 import java.util.List;
 
 public final class IdentityGetResponse extends BaseResponse {
-  private Item item;
+  private ItemStatus item;
   private List<AccountWithOwners> accounts;
 
-  public Item getItem() {
+  public ItemStatus getItem() {
     return item;
   }
