+++ src/main/java/com/plaid/client/response/ItemStatus.java
@@ -1,28 +1,38 @@
 package com.plaid.client.response;
 
+import com.plaid.client.request.common.Product;
+
+import java.util.List;
+
 public final class ItemStatus {
+  private List<Product> availableProducts;
+  private List<Product> billedProducts;
+  private ErrorResponse error;
+  private String institutionId;
+  private String itemId;
+  private String webhook;
 
-  public static final class Transactions {
-    private String lastFailedUpdate;
-    private String lastSuccessfulUpdate;
-
-    public String getLastFailedUpdate() {
-      return lastFailedUpdate;
-    }
-
-    public String getLastSuccessfulUpdate() {
-      return lastSuccessfulUpdate;
-    }
+  public List<Product> getAvailableProducts() {
+    return availableProducts;
   }
 
-  private Transactions transactions;
-  private String lastWebhook;
-
-  public Transactions getTransactions() {
-    return transactions;
+  public List<Product> getBilledProducts() {
+    return billedProducts;
   }
 
-  public String getLastWebhook() {
-    return lastWebhook;
+  public ErrorResponse getError() {
+    return error;
+  }
+
+  public String getInstitutionId() {
+    return institutionId;
+  }
+
+  public String getItemId() {
+    return itemId;
+  }
+
+  public String getWebhook() {
+    return webhook;
   }
 }
