+++ src/main/java/com/plaid/client/response/ItemWebhookUpdateResponse.java
@@ -4,9 +4,9 @@ package com.plaid.client.response;
  * Response from /item/webhook/update endpoint.
  */
 public final class ItemWebhookUpdateResponse extends BaseResponse {
-  private Item item;
+  private ItemStatus item;
 
-  public Item getItem() {
+  public ItemStatus getItem() {
     return item;
   }
 }
